% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_partners}
\alias{cc_partners}
\title{China Customs trading partner names}
\format{
A data frame with 243 rows and 3 variables:
\describe{
\item{code}{trading partner code}
\item{en}{English trading partner name}
\item{zh}{Chinese trading partner name}
}
}
\source{
\url{http://stats.customs.gov.cn/indexEn}
}
\usage{
cc_partners
}
\description{
A dataset containing all countries and regions listed as trading partners by China Customs. Note that the entries are not necessarily independent states. For example, Canary Islands are listed separately from Spain. The list also contains legacy states, such as Serbia and Montenegro, and residual categories (e.g., Other Oceania nes  ).
}
\keyword{datasets}
