% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_regimes}
\alias{cc_regimes}
\title{China Customs customs regime names}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{code}{customs regime code}
\item{en}{English customs regime name}
\item{zh}{Chinese customs regime name}
}
}
\source{
\url{http://stats.customs.gov.cn/indexEn}
}
\usage{
cc_regimes
}
\description{
A dataset containing all customs regime names used by China Customs in English and Chinese, including their codes.
}
\keyword{datasets}
