% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_variable_names}
\alias{cc_variable_names}
\title{China Customs variable names}
\format{
A data frame with 17 rows and 4 variables:
\describe{
\item{clean_name}{name used for harmonization of variable names}
\item{en}{English variable name}
\item{zh}{Chinese variable name}
\item{col_type}{single character denoting the column type, used to generate a column type specification for parsing data}
}
}
\source{
\url{http://stats.customs.gov.cn/indexEn}
}
\usage{
cc_variable_names
}
\description{
A dataset containing all variable names that can appear in data downloaded from Chinese Customs. Variable names are given in English and Chinese to allow harmonizing them. Not all variables will occur in a given query.
}
\details{
Some clean names map from more than one English spelling because China
Customs has changed the wording over time (e.g. "Supplimentary" vs
"Supplementary"), so the table has more rows than there are clean names.
}
\keyword{datasets}
