% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_commodities}
\alias{cc_commodities}
\title{China Customs commodity codes}
\format{
A data frame with 77,320 rows and 4 variables:
\describe{
\item{code}{commodity code}
\item{year}{year}
\item{en}{English commodity description}
\item{zh}{Chinese commodity description}
}
}
\source{
\url{http://stats.customs.gov.cn/indexEn}
}
\usage{
cc_commodities
}
\description{
A dataset containing all commodity descriptions used by China Customs in English and Chinese, including their codes.
}
\details{
This dataset only contains codes at 8-digit level because codes at 2, 4, and 6-digit level are standardized internationally. Use \code{concordance::get_desc()} to retrieve descriptions for those. China Customs further subdivides the 8-digit codes reported in this dataset into 10-digit codes.

Note that the usage of codes varies by year. China Customs does not indicate which HS revision its statistics are based on but a small-n investigation suggests that the latest version is used for each record.
}
\keyword{datasets}
