% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_partners}
\alias{cc_partners}
\title{China Customs trading partner names}
\format{
A data frame with 243 rows and 3 variables:
\describe{
\item{code}{trading partner code}
\item{en}{English trading partner name}
\item{zh}{Chinese trading partner name}
}

A data frame with 51,769 rows and 4 variables:
\describe{
\item{code}{commodity code}
\item{year}{year}
\item{en}{English commodity name}
\item{zh}{Chinese commodity name}
}
}
\source{
\url{43.248.49.97/indexEn}

\url{43.248.49.97/indexEn}
}
\usage{
cc_partners

cc_partners
}
\description{
A dataset containing all countries and regions listed as trading partners by China Customs. Note that the entries are not necessarily independent states. For example, Canary Islands are listed separately from Spain. The list also contains legacy states, such as Serbia and Montenegro, and residual categories (e.g., Other Oceania nes  ).

A dataset containing all 8-digit commodity codes and their descriptions in English and Chinese from China Customs, by year. HS codes up to the 6-digit level are standardized and can be queried through the \code{concordance} package. Note that it is unclear which HS revision China Customs was using in a given year.
}
\keyword{datasets}
