% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_read_csv.R
\name{cc_read_csv}
\alias{cc_read_csv}
\title{Read and Clean Trade Data From China Customs}
\usage{
cc_read_csv(file, drop_descriptions = TRUE)
}
\arguments{
\item{drop_descriptions}{}
}
\description{
Read and Clean Trade Data From China Customs
}
