% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_read_csv.R
\name{cc_read_csv}
\alias{cc_read_csv}
\title{Read and Clean Trade Data From China Customs}
\usage{
cc_read_csv(paths, drop_descriptions = TRUE)
}
\arguments{
\item{drop_descriptions}{Keep only codes for commodity, partner country, province, and customs regime? Default is \code{TRUE} because descriptions are redundant and often include spelling mistakes.}

\item{file}{Path to one or multiple CSV files downloaded from China Customs. Files can be in English, Chinese, or a mix of both.}
}
\value{
A tibble
}
\description{
This is a function to read CSV files downloaded from China Customs with the correct encoding and column type specification, no matter which columns are present in the data, their order and the language. The function harmonizes column names between English and Chinese files and converts dates.
}
\details{
Note that this function does not perform any sanity checks on the data, so make sure no individual CSV file was cut off after 10,000 lines. The function expects (and ignores) a trailing comma in each line of the CSV file which is normally added by China Customs. The order of columns has to be the same across files.
}
