% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryname.R
\name{countryname}
\alias{countryname}
\title{Harmonize (Chinese) Country Names}
\usage{
countryname(sourcevar, origin = "regex", destination = "iso3c")
}
\arguments{
\item{destination}{}
}
\description{
Harmonize (Chinese) Country Names
}
